Began the implementation of command sinks. This affects a lot of the existing Commandment codebase.
Deleted the CommandManager as it is no longer needed.
Moved the UserInputCommander from Graphics to Commandment, where it makes some more sense.
Updated the Commanders so they use a command sink for their command execution pipeline instead of relying on a CommandManager.
Created the skeleton of several sinks (all in Orion.Commandment/Sinks).
Began implementing the CommandSynchronization sink that will act as a mediator between the network and the game. Work in progress; not yet in use.

Obviously updated the main method so everything falls in place.

It all compiles and works.